testeExcluir

+++ pessoas_excluir.php
@@ -7,9 +7,12 @@
 
     
    //echo $idPessoa;
+   if(strlen($idPessoa) > 0){
+      $sql = "DELETE FROM pessoas WHERE id =" .$idPessoa ;
+      mysqli_query($conexao_bd, $sql);
+   }
    //verifico se é vazio:
-   $sql = "DELETE FROM pessoas WHERE id =" .$idPessoa ;
-   mysqli_query($conexao_bd, $sql);
+   
   
    mysqli_close($conexao_bd);
    header("location:pessoas_list.php");
